requirements.txt
numpy==1.13.3
PyAudio==0.2.11
PyYAML==6.0
sounddevice==0.4.3
torch==1.10.0a0+git36449ea
